🎨 Move the `getProductSubscriptionKeys` to the sample app in order to remove BuildConfig reference from the SDK
